mipha v0.0.1 Mipha.Follows

The Follows context.

Functions

can_follow?(clauses)

can_follow?(Keyword.t()) :: boolean()

Returns true if the follower can follow the followable. false otherwise.

change_follow(follow)

Returns an %Ecto.Changeset{} for tracking follow changes.

Examples

iex> change_follow(follow)
%Ecto.Changeset{source: %Follow{}}

create_follow(attrs \\ %{})

Creates a follow.

Examples

iex> create_follow(%{field: value})
{:ok, %Follow{}}

iex> create_follow(%{field: bad_value})
{:error, %Ecto.Changeset{}}

delete_follow(follow)

delete_follow(Mipha.Follows.Follow.t()) ::
  {:ok, Mipha.Follows.Follow.t()} | {:error, Ecto.Changeset.t()}

delete_follow(Keyword.t()) ::
  {:ok, Mipha.Follows.Follow.t()} | {:error, Ecto.Changeset.t()}

Deletes a Follow.

Examples

iex> delete_follow(follow)
{:ok, %Follow{}}

iex> delete_follow(follow)
{:error, %Ecto.Changeset{}}

follow_user(list)

follow_user(Keyword.t()) :: {:ok, Mipha.Follows.Follow.t()} | {:error, any()}

Follow user.

Examples

follow_user(follower: %User{}, user: %User{})
{:ok, follow}

follow_user(follower: %User{}, user: %User{})
{:error, _}

get_follow!(id)

Gets a single follow.

Raises Ecto.NoResultsError if the Follow does not exist.

Examples

iex> get_follow!(123)
%Follow{}

iex> get_follow!(456)
** (Ecto.NoResultsError)

get_follow(id)

get_follow(String.t() | non_neg_integer()) :: Mipha.Follows.Follow.t() | nil

get_follow(Keyword.t()) :: Mipha.Follows.Follow.t() | nil

Gets a single follow.

Examples

iex> get_follow(123)
%Follow{}

iex> get_follow(456)
nil

iex> get_follow(follower_id: 123, user_id: 456)
%Follow{}

iex> get_follow(follower_id: 123, user_id: 458)
nil

get_followee_count(user)

get_followee_count(Mipha.Accounts.User.t()) :: non_neg_integer()

Gets the followee count of a user.

Examples

iex> get_followee_count(%User{}) 42

get_follower_count(clauses)

get_follower_count(Keyword.t()) :: non_neg_integer()

Gets the follower count of a followable.

Examples

iex> get_follower_count([user: %User{}])
4

iex> get_follower_count([artwork: %Artwork{}])
8

iex> get_follower_count([topic: %Topic{}])
12

has_followed?(clauses)

has_followed?(Keyword.t()) :: boolean()

Returns true if the user has followed the followable. false otherwise.

Examples

iex> has_followed?(follower: %User{}, user: %User{})
true

iex> has_followed?(follower: %User{}, topic: %Topic{})
false

insert_follow(attrs \\ %{})

Insert a follow.

Examples

iex> insert_follow(attrs)
{:ok, %Follow{}}

iex> insert_follow(attrs)
{:error, %Ecto.Changeset{}}

list_follows()

Returns the list of follows.

Examples

iex> list_follows()
[%Follow{}, ...]

unfollow_user(list)

unfollow_user(Keyword.t()) :: {:ok, Mipha.Follows.Follow.t()} | {:error, any()}

Unfollow user.

Examples

follow_user(follower: %User{}, user: %User{})
{:ok, follow}

follow_user(follower: %User{}, user: %User{})
{:error, _}

update_follow(follow, attrs)

Updates a follow.

Examples

iex> update_follow(follow, %{field: new_value})
{:ok, %Follow{}}

iex> update_follow(follow, %{field: bad_value})
{:error, %Ecto.Changeset{}}