mipha v0.0.1 MiphaWeb.Router.Helpers

Module with named helpers generated from MiphaWeb.Router.

Functions

admin_company_path(conn_or_endpoint, action)

admin_company_path(conn_or_endpoint, action, params)

admin_company_path(conn_or_endpoint, action, id, params)

admin_company_url(conn_or_endpoint, action)

admin_company_url(conn_or_endpoint, action, params)

admin_company_url(conn_or_endpoint, action, id, params)

admin_node_path(conn_or_endpoint, action)

admin_node_path(conn_or_endpoint, action, params)

admin_node_path(conn_or_endpoint, action, id, params)

admin_node_url(conn_or_endpoint, action)

admin_node_url(conn_or_endpoint, action, params)

admin_node_url(conn_or_endpoint, action, id, params)

admin_notification_path(conn_or_endpoint, action)

admin_notification_path(conn_or_endpoint, action, params)

admin_notification_path(conn_or_endpoint, action, id, params)

admin_notification_url(conn_or_endpoint, action)

admin_notification_url(conn_or_endpoint, action, params)

admin_notification_url(conn_or_endpoint, action, id, params)

admin_page_path(conn_or_endpoint, action)

admin_page_path(conn_or_endpoint, action, params)

admin_page_url(conn_or_endpoint, action)

admin_page_url(conn_or_endpoint, action, params)

admin_reply_path(conn_or_endpoint, action)

admin_reply_path(conn_or_endpoint, action, params)

admin_reply_path(conn_or_endpoint, action, id, params)

admin_reply_url(conn_or_endpoint, action)

admin_reply_url(conn_or_endpoint, action, params)

admin_reply_url(conn_or_endpoint, action, id, params)

admin_team_path(conn_or_endpoint, action)

admin_team_path(conn_or_endpoint, action, params)

admin_team_path(conn_or_endpoint, action, id, params)

admin_team_url(conn_or_endpoint, action)

admin_team_url(conn_or_endpoint, action, params)

admin_team_url(conn_or_endpoint, action, id, params)

admin_topic_path(conn_or_endpoint, action)

admin_topic_path(conn_or_endpoint, action, params)

admin_topic_url(conn_or_endpoint, action)

admin_topic_url(conn_or_endpoint, action, params)

admin_user_path(conn_or_endpoint, action)

admin_user_path(conn_or_endpoint, action, params)

admin_user_path(conn_or_endpoint, action, id, params)

admin_user_url(conn_or_endpoint, action)

admin_user_url(conn_or_endpoint, action, params)

admin_user_url(conn_or_endpoint, action, id, params)

auth_path(conn_or_endpoint, action)

auth_path(conn_or_endpoint, action, provider)

auth_path(conn_or_endpoint, action, provider, params)

auth_url(conn_or_endpoint, action)

auth_url(conn_or_endpoint, action, provider)

auth_url(conn_or_endpoint, action, provider, params)

callback_path(conn_or_endpoint, action)

callback_path(conn_or_endpoint, action, params)

callback_url(conn_or_endpoint, action)

callback_url(conn_or_endpoint, action, params)

company_path(conn_or_endpoint, action)

company_path(conn_or_endpoint, action, params)

company_path(conn_or_endpoint, action, id, params)

company_url(conn_or_endpoint, action)

company_url(conn_or_endpoint, action, params)

company_url(conn_or_endpoint, action, id, params)

join_path(conn_or_endpoint, action)

join_path(conn_or_endpoint, action, params)

join_url(conn_or_endpoint, action)

join_url(conn_or_endpoint, action, params)

location_path(conn_or_endpoint, action)

location_path(conn_or_endpoint, action, params)

location_path(conn_or_endpoint, action, id, params)

location_url(conn_or_endpoint, action)

location_url(conn_or_endpoint, action, params)

location_url(conn_or_endpoint, action, id, params)

logout_path(conn_or_endpoint, action)

logout_path(conn_or_endpoint, action, params)

logout_url(conn_or_endpoint, action)

logout_url(conn_or_endpoint, action, params)

notification_path(conn_or_endpoint, action)

notification_path(conn_or_endpoint, action, params)

notification_url(conn_or_endpoint, action)

notification_url(conn_or_endpoint, action, params)

page_path(conn_or_endpoint, action)

page_path(conn_or_endpoint, action, params)

page_url(conn_or_endpoint, action)

page_url(conn_or_endpoint, action, params)

path(data, path)

Generates the path information including any necessary prefix.

search_path(conn_or_endpoint, action)

search_path(conn_or_endpoint, action, params)

search_url(conn_or_endpoint, action)

search_url(conn_or_endpoint, action, params)

session_path(conn_or_endpoint, action)

session_path(conn_or_endpoint, action, params)

session_url(conn_or_endpoint, action)

session_url(conn_or_endpoint, action, params)

setting_account_path(conn_or_endpoint, action)

setting_account_path(conn_or_endpoint, action, params)

setting_account_url(conn_or_endpoint, action)

setting_account_url(conn_or_endpoint, action, params)

setting_password_path(conn_or_endpoint, action)

setting_password_path(conn_or_endpoint, action, params)

setting_password_url(conn_or_endpoint, action)

setting_password_url(conn_or_endpoint, action, params)

setting_path(conn_or_endpoint, action)

setting_path(conn_or_endpoint, action, params)

setting_profile_path(conn_or_endpoint, action)

setting_profile_path(conn_or_endpoint, action, params)

setting_profile_url(conn_or_endpoint, action)

setting_profile_url(conn_or_endpoint, action, params)

setting_reward_path(conn_or_endpoint, action)

setting_reward_path(conn_or_endpoint, action, params)

setting_reward_url(conn_or_endpoint, action)

setting_reward_url(conn_or_endpoint, action, params)

setting_url(conn_or_endpoint, action)

setting_url(conn_or_endpoint, action, params)

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

team_path(conn_or_endpoint, action)

team_path(conn_or_endpoint, action, params)

team_path(conn_or_endpoint, action, id, params)

team_url(conn_or_endpoint, action)

team_url(conn_or_endpoint, action, params)

team_url(conn_or_endpoint, action, id, params)

topic_path(conn_or_endpoint, action)

topic_path(conn_or_endpoint, action, params)

topic_path(conn_or_endpoint, action, id, params)

topic_reply_path(conn_or_endpoint, action, topic_id)

topic_reply_path(conn_or_endpoint, action, topic_id, params)

topic_reply_path(conn_or_endpoint, action, topic_id, id, params)

topic_reply_reply_path(conn_or_endpoint, action, topic_id, reply_id)

topic_reply_reply_path(conn_or_endpoint, action, topic_id, reply_id, params)

topic_reply_reply_url(conn_or_endpoint, action, topic_id, reply_id)

topic_reply_reply_url(conn_or_endpoint, action, topic_id, reply_id, params)

topic_reply_url(conn_or_endpoint, action, topic_id)

topic_reply_url(conn_or_endpoint, action, topic_id, params)

topic_reply_url(conn_or_endpoint, action, topic_id, id, params)

topic_url(conn_or_endpoint, action)

topic_url(conn_or_endpoint, action, params)

topic_url(conn_or_endpoint, action, id, params)

url(data)

Generates the connection/endpoint base URL without any path information.

user_collections_path(conn_or_endpoint, action, name)

user_collections_path(conn_or_endpoint, action, name, params)

user_collections_url(conn_or_endpoint, action, name)

user_collections_url(conn_or_endpoint, action, name, params)

user_follow_path(conn_or_endpoint, action, name)

user_follow_path(conn_or_endpoint, action, name, params)

user_follow_url(conn_or_endpoint, action, name)

user_follow_url(conn_or_endpoint, action, name, params)

user_followers_path(conn_or_endpoint, action, name)

user_followers_path(conn_or_endpoint, action, name, params)

user_followers_url(conn_or_endpoint, action, name)

user_followers_url(conn_or_endpoint, action, name, params)

user_following_path(conn_or_endpoint, action, name)

user_following_path(conn_or_endpoint, action, name, params)

user_following_url(conn_or_endpoint, action, name)

user_following_url(conn_or_endpoint, action, name, params)

user_path(conn_or_endpoint, action)

user_path(conn_or_endpoint, action, name)

user_path(conn_or_endpoint, action, name, params)

user_replies_path(conn_or_endpoint, action, name)

user_replies_path(conn_or_endpoint, action, name, params)

user_replies_url(conn_or_endpoint, action, name)

user_replies_url(conn_or_endpoint, action, name, params)

user_reward_path(conn_or_endpoint, action, name)

user_reward_path(conn_or_endpoint, action, name, params)

user_reward_url(conn_or_endpoint, action, name)

user_reward_url(conn_or_endpoint, action, name, params)

user_stars_path(conn_or_endpoint, action, name)

user_stars_path(conn_or_endpoint, action, name, params)

user_stars_url(conn_or_endpoint, action, name)

user_stars_url(conn_or_endpoint, action, name, params)

user_topics_path(conn_or_endpoint, action, name)

user_topics_path(conn_or_endpoint, action, name, params)

user_topics_url(conn_or_endpoint, action, name)

user_topics_url(conn_or_endpoint, action, name, params)

user_unfollow_path(conn_or_endpoint, action, name)

user_unfollow_path(conn_or_endpoint, action, name, params)

user_unfollow_url(conn_or_endpoint, action, name)

user_unfollow_url(conn_or_endpoint, action, name, params)

user_url(conn_or_endpoint, action)

user_url(conn_or_endpoint, action, name)

user_url(conn_or_endpoint, action, name, params)